Refactor LoggingProducerContext to a common location
